feat(recipe-form): implement recipe ingredients, labels, main image, and user tagging sections with search functionality
