+++ src/modules/recipes/components/index.ts
@@ -6,3 +6,6 @@ export { RecipeDetailView } from './recipe-detail-view';
 export { RecipeForm } from './recipe-form';
 export { RecipeRating } from './recipe-rating';
 export { UserRecipesList } from './user-recipes-list';
+
+// Recipe form sub-components (for advanced usage)
+export * from './recipe-form/types';
