+++ src/modules/recipes/hooks/index.ts
@@ -1,4 +1,9 @@
 export * from './use-comment-manager';
+export * from './use-cooking-steps';
+export * from './use-image-upload';
+export * from './use-ingredient-search';
+export * from './use-label-search';
 export * from './use-recipe-actions';
 export * from './use-signalr-connection';
 export * from './use-user-recipes';
+export * from './use-user-tagging';
